virtio: add vhost-user-vsock-pci device
Based on upstream 9b83bb27477f3fb2c08db32ea2ee9e8f817f0f9d
